# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ELFOR LED Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  imageUrl?: string;
  sortOrder?: number;
}

export interface CategoryPatch {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
}

export interface PriceTier {
  minQty: number;
  price: number;
}

export interface ProductSpec {
  key: string;
  value: string;
  /** @nullable */
  unit?: string | null;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  /** @nullable */
  series?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  priceTiers?: PriceTier[];
  specs?: ProductSpec[];
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  /** @nullable */
  power?: string | null;
  /** @nullable */
  lumens?: number | null;
  /** @nullable */
  colorTemp?: string | null;
  /** @nullable */
  ipRating?: string | null;
  /** @nullable */
  warranty?: string | null;
  stock: number;
  featured?: boolean;
  inStock?: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  series?: string;
  categoryId: number;
  shortDescription?: string;
  fullDescription?: string;
  price: number;
  oldPrice?: number;
  priceTiers?: PriceTier[];
  specs?: ProductSpec[];
  imageUrl?: string;
  images?: string[];
  power?: string;
  lumens?: number;
  colorTemp?: string;
  ipRating?: string;
  warranty?: string;
  stock: number;
  featured?: boolean;
}

export interface ProductPatch {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  sku?: string;
  /** @nullable */
  series?: string | null;
  categoryId?: number;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  price?: number;
  /** @nullable */
  oldPrice?: number | null;
  priceTiers?: PriceTier[];
  specs?: ProductSpec[];
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  /** @nullable */
  power?: string | null;
  /** @nullable */
  lumens?: number | null;
  /** @nullable */
  colorTemp?: string | null;
  /** @nullable */
  ipRating?: string | null;
  /** @nullable */
  warranty?: string | null;
  stock?: number;
  featured?: boolean;
  inStock?: boolean;
}

export type ArticleType = typeof ArticleType[keyof typeof ArticleType];


export const ArticleType = {
  news: 'news',
  article: 'article',
} as const;

export interface Article {
  id: number;
  title: string;
  slug: string;
  type: ArticleType;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type ArticleInputType = typeof ArticleInputType[keyof typeof ArticleInputType];


export const ArticleInputType = {
  news: 'news',
  article: 'article',
} as const;

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  type: ArticleInputType;
  excerpt?: string;
  content?: string;
  imageUrl?: string;
  published?: boolean;
}

export type ArticlePatchType = typeof ArticlePatchType[keyof typeof ArticlePatchType];


export const ArticlePatchType = {
  news: 'news',
  article: 'article',
} as const;

export interface ArticlePatch {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  slug?: string;
  type?: ArticlePatchType;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  published?: boolean;
}

export interface OrderItem {
  productId: number;
  productName: string;
  productSku: string;
  quantity: number;
  unitPrice: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  /** @nullable */
  customerCompany?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  comment?: string | null;
  status: OrderStatus;
  totalAmount: number;
  items: OrderItem[];
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  customerEmail: string;
  customerCompany?: string;
  deliveryAddress?: string;
  comment?: string;
  items: OrderItem[];
}

export type OrderPatchStatus = typeof OrderPatchStatus[keyof typeof OrderPatchStatus];


export const OrderPatchStatus = {
  new: 'new',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderPatch {
  status?: OrderPatchStatus;
  /** @nullable */
  comment?: string | null;
}

export type DashboardStatsOrdersByStatusItem = {
  status: string;
  count: number;
};

export interface DashboardStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  newOrdersCount: number;
  totalCategories: number;
  totalArticles: number;
  recentOrders: Order[];
  ordersByStatus: DashboardStatsOrdersByStatusItem[];
}

export type ListProductsParams = {
categoryId?: number;
featured?: boolean;
search?: string;
limit?: number;
offset?: number;
};

export type ListArticlesParams = {
type?: ListArticlesType;
published?: boolean;
limit?: number;
offset?: number;
};

export type ListArticlesType = typeof ListArticlesType[keyof typeof ListArticlesType];


export const ListArticlesType = {
  news: 'news',
  article: 'article',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
limit?: number;
offset?: number;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  new: 'new',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;