# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ELFOR LED Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ===== PRODUCTS =====
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles/news
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [news, article]
        - name: published
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create an article/news (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get an article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticlePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ===== ORDERS =====
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, processing, shipped, delivered, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  # ===== STATS =====
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
    CategoryPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer

    # ===== PRODUCT =====
...
    Article:
      type: object
      required: [id, title, slug, type, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [news, article]
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ArticleInput:
      type: object
      required: [title, slug, type]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        type:
          type: string
          enum: [news, article]
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        published:
          type: boolean
    ArticlePatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        type:
          type: string
          enum: [news, article]
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean

    # ===== ORDER =====
    OrderItem:
      type: object
      required: [productId, productName, productSku, quantity, unitPrice]
      properties:
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
    Order:
      type: object
      required: [id, customerName, customerPhone, customerEmail, status, totalAmount, items, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        customerCompany:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, processing, shipped, delivered, cancelled]
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, customerPhone, customerEmail, items]
      properties:
        customerName:
          type: string
          minLength: 1
        customerPhone:
          type: string
          minLength: 1
        customerEmail:
          type: string
          minLength: 1
        customerCompany:
          type: string
        deliveryAddress:
          type: string
        comment:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderPatch:
      type: object
      properties:
        status:
          type: string
          enum: [new, processing, shipped, delivered, cancelled]
        comment:
          type: ["string", "null"]

    # ===== STATS =====
...