# lib/api-zod/src/generated/types/articlePatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ELFOR LED Store API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticlePatchType } from './articlePatchType';

export interface ArticlePatch {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  slug?: string;
  type?: ArticlePatchType;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  published?: boolean;
}