# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ELFOR LED Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 1 */
  customerEmail: string;
  customerCompany?: string;
  deliveryAddress?: string;
  comment?: string;
  items: OrderItem[];
}