# lib/api-zod/src/generated/types/orderPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ELFOR LED Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPatchStatus } from './orderPatchStatus';

export interface OrderPatch {
  status?: OrderPatchStatus;
  /** @nullable */
  comment?: string | null;
}